Revert "springboot 3.3.3"
This reverts commit 48329791fff38a65c39ee52ccf0ad279d04f2c57.

+++ cratos-manage/src/main/java/com/baiyi/cratos/facade/impl/BusinessCredentialFacadeImpl.java
@@ -45,7 +45,9 @@ public class BusinessCredentialFacadeImpl implements BusinessCredentialFacade {
                 .businessId(business.getBusinessId())
                 .build();
         BusinessCredential businessCredential = businessCredentialService.getByUniqueKey(query);
-        businessCredentialService.deleteById(businessCredential.getId());
+        if (businessCredential != null) {
+            businessCredentialService.deleteById(businessCredential.getId());
+        }
     }
 
     @Override
@@ -55,7 +57,9 @@ public class BusinessCredentialFacadeImpl implements BusinessCredentialFacade {
                 .businessType(business.getBusinessType())
                 .businessId(business.getBusinessId())
                 .build();
-        businessCredentialService.add(businessCredential);
+        if (businessCredentialService.getByUniqueKey(businessCredential) == null) {
+            businessCredentialService.add(businessCredential);
+        }
     }
 
     @Override
